WIP on controller. Not quite working
