+++ jumpstreet/__init__.py
@@ -1,1 +1,1 @@
-from . import broker, context, detection, display, frontend, muxer, tracking, trigger, utils
+from . import broker, context, controllers, detection, display, frontend, muxer, tracking, trigger, utils
